Source PDF and Output Images
For the Capstone project for DataScience Masters Course ," for the given Problem Statement i.e identifying best model to prevent churning of customers and suggesting retension Programs" , the various images of the output while preparing the Dataset and training the model are uplaoded. The Conclusions made are uplaoded in Conclusion1, and Conclusion2 file . The Source PDF file is also uplaoded , which contains the  code for solving the problem statement.